v1.9 — home at boot, companion URL is read-only, and a dead-code audit
HOME AT BOOT

autoHome defaults to true, 'a' sets it and 'A' reports it, and a REAL module with the
flag set drives its reel to the blank flap on power-up. This emulation stored the flag,
reported it over the protocol, and then quietly ignored it -- on the reasoning that a
drawn reel has no position to lose.

True. But it does have a position to REMEMBER, and /vmods.dat faithfully restored it, so
a reboot brought the wall up still showing whatever half-finished sentence was on it when
the power went. That is not "no position to lose", it is stale content presented as
current. The flag is honoured now, and the wall mirror is seeded from the homed reels so
/api/display/state does not report "?" for cells we just homed ourselves.

THE COMPANION URL IS NOT OURS TO SET

The companion registers it. The Settings card that let you type it in is gone, along with
saveCompanion/clearCompanion/loadCompanion, and /api/status now carries the URL so the
Status tab can show it read-only -- with a line saying so, because someone will go looking
for the field they remember.

DEAD CODE, after everything that has been removed

  - the RS-485 serial parameters: baud, data bits, parity, stop bits. There is no UART.
    The ONLY reader of cfg.rs485Baud was a debug line that called it "cosmetic" out loud.
    With them go POST /api/config/rs485 and the "baud" field in /api/config and /api/status.
  - /api/flap/version + sfSendVersionAndWait
  - /api/flap/all + sfSendAndCaptureDump + the gDump capture mailbox + struct DumpCapture
  - /api/flap/identify (the boot reconciliation self-heals the registry now)
  - the 'A' reply's flap-set tail, which was parsed and then handed to nobody

Every one of them existed to query the emulated bus for a COMPILE-TIME CONSTANT: the
version is VM_FW_VERSION, the calibration is nominal, the flap set is the one shared reel.
The gateway was asking the bus to tell it things it already knew.

Verified empirically before cutting, not assumed: the handlers were instrumented and the
board watched for 75 seconds with the companion actively driving the wall. Not one of them
fired. (The companion drives entirely through /api/display/cells now -- it has already
moved off the legacy batch API.)

-Wall IS ON NOW, AND THE BUILD IS WARNING-CLEAN

Turning it on immediately found a REAL LATENT NULL DEREFERENCE:

    SFModule* m = sfUpsert(id, NULL);                          // modules.cpp
    ...
    SFModule* m = (id != 255) ? sfFindById(id) : sfFindBySN(sn);   // sn is NULL here
    ...
    if (strcmp(sfModules[i].serialNum, sn) == 0)                   // strcmp(x, NULL)

Nothing on this firmware ever produces id 255 -- the virtual modules take their cell index
-- so the only thing between that call and a crash was an invariant nobody had written
down. Now there is a guard.

Also cleared: 8 format specifiers printing uint32_t (which is `long unsigned int` here)
with %u/%X -- latent UB that happens to work because both are 32 bits; three C++20
volatile-increment deprecations; and two format-truncation warnings on log lines that are
bounded ON PURPOSE, now said so with a precision instead of left for the compiler to guess.

RAM 30.1% -> 28.7%. Flash 64.6% -> 64.1%. Zero warnings.

+++ src/ota.cpp
@@ -199,7 +199,7 @@ void handleOTAUpload() {
     if (WiFi.status() == WL_CONNECTED) {
       wifiSetApActive(false);   // drop fallback AP if it happens to be up
       WiFi.setSleep(false);
-      printf("[OTA] AP down + modem sleep off for upload (heap=%u)\n", ESP.getFreeHeap());
+      printf("[OTA] AP down + modem sleep off for upload (heap=%u)\n", (unsigned)ESP.getFreeHeap());
     }
     // UPDATE_SIZE_UNKNOWN lets the Update library size the partition itself.
     if (!Update.begin(UPDATE_SIZE_UNKNOWN)) {
